Rename slacksocketmode to socketmode
