+++ socketmode/deadman.go
@@ -1,4 +1,4 @@
-package slacksocketmode
+package socketmode
 
 import "time"
