+++ socketmode/socket_mode_managed_conn.go
@@ -1,4 +1,4 @@
-package slacksocketmode
+package socketmode
 
 import (
 	"bytes"
